Move Kinesis to core_ops to reduce the number of .so
