+++ tensorflow_io/kinesis/BUILD
@@ -7,18 +7,21 @@ load(
     "tf_io_copts",
 )
 
-cc_binary(
-    name = "python/ops/_kinesis_ops.so",
+cc_library(
+    name = "kinesis_ops",
     srcs = [
+        "kernels/aws_kernels.cc",
         "kernels/kinesis_dataset_ops.cc",
         "ops/dataset_ops.cc",
     ],
+    hdrs = [
+        "kernels/aws_kernels.h",
+    ],
     copts = tf_io_copts(),
-    linkshared = 1,
+    linkstatic = True,
     deps = [
+        "//tensorflow_io/core:dataset_ops",
         "@aws",
         "@curl",
-        "@local_config_tf//:libtensorflow_framework",
-        "@local_config_tf//:tf_header_lib",
     ],
 )
